requirements.txt
flask==2.3.3
flask-sqlalchemy==2.5.1
flask-login==0.6.6
flask-migrate==4.0.4
flask-wtf==1.1.1
wtforms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
openai>=1.0.0
requests==2.31.0
werkzeug==2.3.7
gunicorn==21.2.0
# Заменяем psycopg2-binary на pg8000 для совместимости с Python 3.13
pg8000==1.30.1
alembic==1.12.0
SQLAlchemy==1.4.50
pytest==7.4.3
stripe==6.0.0
pillow==10.0.0
unidecode==1.3.6
python-dateutil==2.8.2
slugify==0.0.1
